+++ src/main/java/com/example/airbnb/service/RoomServiceImpl.java
@@ -1,15 +1,18 @@
 package com.example.airbnb.service;
 
 import com.example.airbnb.Exceptions.ResourceNotFoundException;
+import com.example.airbnb.Exceptions.UnAuthorizedException;
 import com.example.airbnb.Mapper.RoomMapper;
 import com.example.airbnb.dto.RoomDTO;
 import com.example.airbnb.entity.Hotel;
 import com.example.airbnb.entity.Room;
+import com.example.airbnb.entity.User;
 import com.example.airbnb.repository.HotelRepository;
 import com.example.airbnb.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -26,10 +29,10 @@ public class RoomServiceImpl implements RoomService {
     private final InventoryServiceImpl inventoryService;
 
     public RoomDTO FindRoomById(Long id) {
-        Room room = roomRepository.findById(id).orElseThrow(
-                () -> new ResourceNotFoundException("Room not found with id " + id));
+        Room room = roomRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Room not found with id " + id));
         return roomMapper.toRoomDTO(room);
     }
+
     @Override
     public RoomDTO getRoomById(Long roomId) {
         return null;
@@ -38,23 +41,24 @@ public class RoomServiceImpl implements RoomService {
     @Override
     @Transactional
     public RoomDTO CreateRoom(Long hotelId, RoomDTO roomDTO) {
-        log.info("Creating room with hotelId {} and roomId {}", hotelId , roomDTO);
-        Hotel hotel = hotelRepository.findById(hotelId)
-                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id " + hotelId));
+        log.info("Creating room with hotelId {} and roomId {}", hotelId, roomDTO);
+        Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id " + hotelId));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizedException("User is not owner of this hotel with id " + user.getId());
+        }
         Room room = roomMapper.toRoom(roomDTO);
         room.setHotel(hotel);
 
         Room savedRoom = roomRepository.save(room);
 
-       if(hotel.getActive()){
-           inventoryService.initializeRoomForYear(savedRoom);
-       }
+        if (hotel.getActive()) {
+            inventoryService.initializeRoomForYear(savedRoom);
+        }
         return roomMapper.toRoomDTO(savedRoom);
     }
-
-
-
 
     public RoomDTO UpdateRoom(RoomDTO roomDTO) {
         Room room = roomMapper.toRoom(roomDTO);
@@ -63,33 +67,27 @@ public class RoomServiceImpl implements RoomService {
         return roomMapper.toRoomDTO(savedRoom);
     }
 
-
-    public void isRoomExist(Long id) {
-        boolean isExist = roomRepository.existsRoomById(id);
-        if (!isExist) {
-            throw new ResourceNotFoundException("Room not found with id " + id);
-        }
-    }
-
     @Override
     public List<RoomDTO> getAllRoomsInHotel(Long hotelId) {
         log.info("Getting all rooms in hotel with id {}", hotelId);
-        Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(
-                () ->  new ResourceNotFoundException("Hotel not found with id " + hotelId));
+        Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id " + hotelId));
+
         List<Room> rooms = hotel.getRooms();
         return rooms.stream().map(roomMapper::toRoomDTO).collect(Collectors.toList());
     }
-
-
 
 
     @Override
     @Transactional
     public void deleteRoomById(Long roomId) {
         log.info("Deleting room with id {}", roomId);
-       Room room = roomRepository.findById(roomId).orElseThrow(
-               () ->  new ResourceNotFoundException("Room not found with id " + roomId)
-       );
+        Room room = roomRepository.findById(roomId).orElseThrow(() -> new ResourceNotFoundException("Room not found with id " + roomId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.equals(room.getHotel().getOwner())) {
+            throw new UnAuthorizedException("User is not owner of this room with id " + user.getId());
+        }
         inventoryService.deleteAllByRoom(room);
         roomRepository.deleteById(roomId);
